# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Stars Decoded Natal Chart Report API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invites:
    get:
      operationId: listInvites
      tags: [invites]
      summary: List active invites for a profile owned by the current viewer
      description: |
        Returns all unclaimed, unexpired invite tokens for the given profile.
        The caller must own the profile (i.e. be the owner by userId or sessionId).
        Ownership is the effective gate — all active invites for the profile are
        returned regardless of which session originally created them.
      parameters:
        - name: profileId
          in: query
          required: true
          schema:
            type: string
          description: Filter invites to those belonging to this profile
      responses:
        "200":
          description: List of active invite records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InviteRecord"
        "403":
          description: Not the owner of the profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createInvite
      tags: [invites]
      summary: Create an invite token and send the claim email
      description: Generates a single-use signed token tied to a profile (and optional relationship) and emails it to the recipient.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteBody"
      responses:
        "201":
          description: Invite created and email queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteSummary"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Caller is not the owner of the profile/relationship
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Profile is already claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invites/{token}:
    get:
      operationId: getInvite
      tags: [invites]
      summary: Resolve an invite token to its preview metadata
      description: |
        Public endpoint used by the claim landing page to render the invite
        without requiring sign-in. A token that has already been claimed is
        still resolvable (so the landing page can show a "this invite was
        already accepted" state) and is signaled by `alreadyClaimed: true`
        in the response body. Tokens that do not exist or have expired
        return 404.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite preview (may be already-claimed; check `alreadyClaimed`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitePreview"
        "404":
          description: Token not found or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invites/{token}/claim:
    post:
      operationId: claimInvite
      tags: [invites]
      summary: Claim an invite as the signed-in user
      description: Requires Clerk authentication. Marks the token claimed, sets profile.claimed_by_user_id, and adds the user as a participant on the related relationship.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteClaimResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Token not found or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    InviteRecord:
      type: object
      description: A single invite token row as visible to the sender
      properties:
        id: { type: string }
        email: { type: string }
        profileId: { type: string }
        relationshipId: { type: string, nullable: true }
        sentAt: { type: string, description: "ISO-8601 timestamp when the invite was created" }
        expiresAt: { type: string }
        emailDelivered: { type: boolean, nullable: true, description: "null means unknown (pre-migration)" }
        claimedAt: { type: string, nullable: true }
      required: [id, email, profileId, sentAt, expiresAt]
    CreateInviteBody:
      type: object
      properties:
        profileId: { type: string }
        relationshipId: { type: string, nullable: true }
        email: { type: string, format: email }
      required: [profileId, email]
    InviteSummary:
      type: object
      properties:
        id: { type: string }
        token: { type: string }
        email: { type: string }
        profileId: { type: string }
        relationshipId: { type: string, nullable: true }
        expiresAt: { type: string }
        claimUrl: { type: string }
        emailDelivered: { type: boolean }
      required: [id, token, email, profileId, expiresAt, claimUrl, emailDelivered]
    InvitePreview:
      type: object
      properties:
        token: { type: string }
        email: { type: string }
        inviterName: { type: string, nullable: true }
        profileName: { type: string }
        relationshipId: { type: string, nullable: true }
        relationshipReportId: { type: string, nullable: true }
        expiresAt: { type: string }
        alreadyClaimed: { type: boolean }
      required: [token, email, profileName, expiresAt, alreadyClaimed]
    InviteClaimResponse:
      type: object
      properties:
        profileId: { type: string }
        relationshipId: { type: string, nullable: true }
        relationshipReportId: { type: string, nullable: true }
        redirectTo: { type: string }
      required: [profileId, redirectTo]
...